Added support for aliased parent tables in join
Had a problem with a mjoin on an aliased table. These changes fixed my problem.

To test this I used the upload-many.php example. Rename the users table in the database to user and replace the line 22 from:
Editor::inst( $db, 'users' )
to
Editor::inst( $db, 'user as users' )

Without this change this was not working.