Fix: Compatibility with PHP 5.x in strict mode
`abstract static` would cause a warning in PHP 5.x strict mode, even
although it is actually technically correct to use it. While fixed in
PHP 7, we want to maintain compability with 5.4 until there is a really
strong reason to drop it (having a function body here is not a strong
enough reason).

See also https://stackoverflow.com/questions/999066